refactor: share parseSql helper
